+++ src/interval.h
@@ -27,15 +27,15 @@ public:
 inline Interval Sin(const Interval &i) {
   Float sinLow = std::sin(i.low), sinHigh = std::sin(i.high);
   if (sinLow > sinHigh) std::swap(sinLow, sinHigh);
-  if (i.low < M_PI / 2 && i.high > M_PI / 2) sinHigh = 1.;
-  if (i.low < (3.f / 2.f) * M_PI && i.high > (3.f / 2.f) * M_PI) sinLow = -1.;
+  if (i.low < static_cast<Float>(M_PI) / 2 && i.high > static_cast<Float>(M_PI) / 2) sinHigh = 1.;
+  if (i.low < (3.f / 2.f) * static_cast<Float>(M_PI) && i.high > (3.f / 2.f) * static_cast<Float>(M_PI)) sinLow = -1.;
   return Interval(sinLow, sinHigh);
 }
 
 inline Interval Cos(const Interval &i) {
   Float cosLow = std::cos(i.low), cosHigh = std::cos(i.high);
   if (cosLow > cosHigh) std::swap(cosLow, cosHigh);
-  if (i.low < M_PI && i.high > M_PI) cosLow = -1.;
+  if (i.low < static_cast<Float>(M_PI) && i.high > static_cast<Float>(M_PI)) cosLow = -1.;
   return Interval(cosLow, cosHigh);
 }
