+++ src/rng.cpp
@@ -36,7 +36,7 @@ vec3f random_gen::random_cosine_direction() {
   Float r1 = unif_rand();
   Float r2 = unif_rand();
   Float z = std::sqrt(static_cast<Float>(1)-r2);
-  Float phi = 2.0 * M_PI * r1;
+  Float phi = 2.0 * static_cast<Float>(M_PI) * r1;
   Float x = cos(phi) * std::sqrt(r2);
   Float y = sin(phi) * std::sqrt(r2);
   return(vec3f(x, y, z));
@@ -49,7 +49,7 @@ vec3f random_gen::random_to_sphere(Float radius, Float distance_squared) {
   Float r1 = unif_rand();
   Float r2 = unif_rand();
   Float z = 1.0 + r2 * (std::sqrt(static_cast<Float>(1)-radius * radius / distance_squared) - 1);
-  Float phi = 2.0 * M_PI * r1;
+  Float phi = 2.0 * static_cast<Float>(M_PI) * r1;
   Float x = std::cos(phi) * std::sqrt(1-z*z);
   Float y = std::sin(phi) * std::sqrt(1-z*z);
   return(vec3f(x,y,z));
